@@ -14,7 +14,7 @@
   },
   "devDependencies": {
     "@ulixee/repo-tools": "^1.0.33",
-    "typescript": "^5.3.3"
+    "typescript": "^5.6.2"
   },
   "workspaces": [
     "analyze",
